# internal/handler/v1/response.go
package v1

import (
	"github.com/gin-gonic/gin"
	"github.com/sirupsen/logrus"
)

type errorResponse struct {
	Message string `json:"message"`
}

func newErrorResponse(c *gin.Context, statusCode int, err string) {
	logrus.Error(err)
	c.AbortWithStatusJSON(statusCode, errorResponse{err})
}